+++ source/Qt/application.cpp
