+++ source/Qt/application.h
