+++ source/Qt/asyncimageprovider.h
