+++ source/Qt/imageprovider.cpp
