+++ source/Qt/imageprovider.h
