+++ source/Qt/utils/about_compilation.h
