+++ source/Qt/utils/utils.h
